app/requirements.txt
flask
pymysql
requests
python-etcd
cryptography
redis
pytest
pytest-flask
etcd3

